+++ src/sample_inverted_index.cpp
@@ -6,7 +6,7 @@
 #include "CLI/CLI.hpp"
 #include "binary_freq_collection.hpp"
 #include "invert.hpp"
-#include "util/index_build_utils.hpp"
+#include "util/inverted_index_utils.hpp"
 #include "util/progress.hpp"
 
 using namespace pisa;
